requirements.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-modeltranslation==0.19.9
pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
python-gettext==5.0
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.2
